pyproject.toml
[project]
name = "langchain-gemini-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.3",
    "langchain-google-genai>=4.0.0",
    "python-dotenv>=1.2.1",
    "google-generativeai>=0.3.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "pypdf>=6.4.1",
    "langchain-tavily>=0.2.13",
    "tavily-python>=0.7.14",
]
